Delete, create and update article
Linked (almost?) every article function with the login. You cannot modify/delete/post unless you are logged in. You cannot delete/modify articles that are not yours.

Added the logout option.

Updated the navigation bar to reflect these changes.

Added the edit pages. It does not update the checkpox from 1 to 0 right now, it has to be corrected!!